Rewrite migration system
