+++ .gitignore
@@ -10,3 +10,4 @@ public
 
 docker-compose.override.yml
 lib/models/old
+scripts
