+++ lib/workers/RefreshCharacter.js
@@ -235,7 +235,7 @@ class RefreshCharacterSkillTree extends Interactor {
 
                 let skills = character.get('skills');
                 for (let skill of skills) {
-                    if (skill.skill.toString() == skillMapping[skillId].toString()) {
+                    if (skillMapping[skillId] !== undefined && skill.skill.toString() == skillMapping[skillId].toString()) {
                         if (!skill.queued) {
                             skill.queued = [];
                         }
@@ -298,7 +298,7 @@ class RefreshCharacter extends Interactor {
         try {
             yield [
                 RefreshCharacterStats.run(this),
-                RefreshCharacterAccounts.run(this)
+                RefreshCharacterAccounts.run(this),
             ];
 
             yield RefreshCharacterSkillTree.run(this);
